Stop prematurely expanding rule variables
When running `mk test`, the environment has $target, $prereq, etc from
the outside execution.  Previously we would expand the rules in this
environment on load, incorrectly expanding these variables with the
values from the outside test run.

Now we expand vars when evaluating the rule, once the correct values
for the current evaluation have been passed down in `vars`.

Aside, test14 suffers from the same premature eval, but since the
test is generating the rules via a shell invocation, it seems
reasonable to allow the expansion here -- instead we simply fix the
quoting in the generating shell code.